Moved actor type to domain
