+++ libs/dh/market-participant/actors/domain/project.json
@@ -1,0 +1,8 @@
+{
+  "name": "dh-market-participant-actors-domain",
+  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "sourceRoot": "libs/dh/market-participant/actors/domain",
+  "prefix": "dh",
+  "tags": ["product:dh", "domain:market-participant", "type:domain"]
+}
