+++ libs/dh/market-participant/actors/domain/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/dh-actor';
